Split browser pipeline models
